+++ schemacrawler-api/src/main/java/schemacrawler/schema/IdentifiersBuilder.java
@@ -11,21 +11,23 @@ package schemacrawler.schema;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static java.util.Objects.requireNonNull;
 import static us.fatehi.utility.CollectionsUtility.splitList;
-import static us.fatehi.utility.Utility.isBlank;
-import static us.fatehi.utility.Utility.trimToEmpty;
 
 import java.io.BufferedReader;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
+import static java.util.Objects.requireNonNull;
+
+import static us.fatehi.utility.Utility.isBlank;
+import static us.fatehi.utility.Utility.trimToEmpty;
+
 import schemacrawler.schemacrawler.OptionsBuilder;
 import schemacrawler.schemacrawler.exceptions.InternalRuntimeException;
 import us.fatehi.utility.ioresource.ClasspathInputResource;
@@ -89,7 +91,7 @@ public class IdentifiersBuilder implements OptionsBuilder<IdentifiersBuilder, Id
       }
     }
     // Create a lighter-weight data structure that is serializable
-    return new ArrayList<>(upperCaseWords);
+    return upperCaseWords;
   }
 
   final Collection<String> reservedWords;